+++ src/headers/bitmap.h
@@ -13,7 +13,7 @@ class Bitmap
     Bitmap() : mNumRows{0}, mNumCols{0} {}
     Bitmap(const int32_t cNumRows, const int32_t cNumCols, const lg::Color& crDefaultColor=lg::Black);
     ~Bitmap();
-    void constructBitmap(const std::vector<Vector2<int32_t>>& crPoints, const lg::Color& crColor);
+    void constructBitmap(const std::vector<Vector2<float>>& crPoints, const lg::Color& crColor);
     void fillColor(const int32_t cStartingPoint, const lg::Color& crColor);
     std::vector<std::shared_ptr<Rect>> getBitmap();
   private:
